% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{allFunctions}
\alias{allFunctions}
\title{create and open a file with the content of all the functions in all my packages}
\usage{
allFunctions(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{combineFiles} like quiet=TRUE}
}
\value{
file path
}
\description{
create and open a file with the content of all the functions in all my packages
}
\examples{
# allFunctions()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Okt 2016
}

