% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCodeLink.R
\name{checkCodeLink}
\alias{checkCodeLink}
\title{check for code{} and link{} statements}
\usage{
checkCodeLink(path = ".")
}
\arguments{
\item{path}{Package path, Character string. DEFAULT: "." (current getwd)}
}
\description{
check for code{} and link{} statements where the leading slash 
has been forgotten in the source code
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2019
}
