% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{checkImports}
\alias{checkImports}
\title{Check imports in source code}
\usage{
checkImports(path = ".", filename = "checkImports.R")
}
\arguments{
\item{path}{Package path, Character string. DEFAULT: "." (current getwd)}

\item{filename}{Output file name. Text will be added, previous content remains.
DEFAULT: "checkImports.R"}
}
\description{
For each file in package/R, check whether all used functions are imported
}
\examples{
\dontrun{ # exclude from checks
library(berryFunctions)
setwd(paste0(pathFinder(), "/berryFunctions"))
imps <- checkImports()
openFile("checkImports.R")
for(n in names(imps)[!sapply(imps, is.null)]) openFile(n) ; rm(n)
unlink("checkImports.R")
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
