% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOutdated.R
\name{checkOutdated}
\alias{checkOutdated}
\title{check if installed package is outdated.}
\usage{
checkOutdated(package, path = "C:/Dropbox/Rpack", quiet = FALSE)
}
\arguments{
\item{package}{Package name}

\item{path}{Path containing package folder. DEFAULT: "C:/Dropbox/Rpack"}

\item{quiet}{suppress messaging.}
}
\description{
check if installed package is outdated.
}
