% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{codeSpaces}
\alias{codeSpaces}
\title{Correct spaces in source code}
\usage{
codeSpaces(path = ".")
}
\arguments{
\item{path}{Package path, Character string. DEFAULT: "." (current getwd)}
}
\description{
Remove trailing spaces in source code
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
