% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{convert_crlf_for_git_on_mac}
\alias{convert_crlf_for_git_on_mac}
\title{convert CRLF for git on Mac OS}
\usage{
convert_crlf_for_git_on_mac(folder = ".", infile = NULL, ...)
}
\arguments{
\item{folder}{Folder in which all files shall be processed 
(unless file is given, then \code{folder} is ignored)}

\item{infile}{File name}

\item{\dots}{Arguments passed to \code{\link{dir}}}
}
\description{
convert CRLF for git on Mac OS
}
\seealso{
\url{https://www.reddit.com/r/RStudio/comments/jxx226/git_line_endings_in_r_studio_is_anyone_else/gcz8k31}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2021
}
\keyword{file}
