% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{course}
\alias{course}
\title{open course PDFs}
\usage{
course(tut = TRUE)
}
\arguments{
\item{tut}{Logical: also open Tutorial pdf? DEFAULT: TRUE}
}
\description{
open 4 course PDFs in default viewer
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2019, Sept 2026
}
