% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crancheck.R
\name{crancheck}
\alias{crancheck}
\title{check packages for cran release}
\usage{
crancheck()
}
\value{
nothing
}
\description{
convenience function to run cran package checks on rhub and win-builder
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, included here Aug 2023
}
\keyword{package}
