% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_ds.R
\name{del_ds}
\alias{del_ds}
\title{delete .DS_Store file on Mac OS}
\usage{
del_ds(folder = ".")
}
\arguments{
\item{folder}{Folder in which all (incl recursive) .DS_Store files shall be deleted}
}
\description{
delete .DS_Store file on Mac OS
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sept 2021
}
