% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAll.R
\name{detachAll}
\alias{detachAll}
\title{Detach all packages on search path}
\usage{
detachAll()
}
\description{
Detach all packages on search path, as found by
             \code{\link{sessionInfo}()$otherPkgs}.
             It puts Berry's packages in descending dependency order.
             It tries \code{\link{unloadNamespace}} and 
             \code{\link{detach}(package:pkg, unload=TRUE)}.
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sept 2016
}
