% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installA.R
\name{installA}
\alias{installA}
\title{run installB on all my packages.}
\usage{
installA(path = "C:/Dropbox/Rpack", quiet = TRUE, ...)
}
\arguments{
\item{path}{Path containing package folder. DEFAULT: "C:/Dropbox/Rpack"}

\item{quiet}{Suppress messages? DEFAULT: FALSE}

\item{\ldots}{passed to \code{\link{installB}}}
}
\description{
run installB on all my packages.
}
