% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{installB}
\alias{installB}
\title{install package from local drive}
\usage{
installB(package = "berryFunctions", path = "S:/Dropbox/Public")
}
\arguments{
\item{package}{Package name. DEFAULT: "berryFunctions"}

\item{path}{Path containing package folder. DEFAULT: "S:/Dropbox/Public"}
}
\description{
Basically just an enhanced shortcut to \code{devtools::\link[devtools]{install}}
}
\details{
Removes function objects from workspace and (toDo!) unloads reverse dependencies first.
}
\examples{
# installB()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2014
}
\seealso{
\code{\link{loadPackages}}
}
\keyword{package}

