\name{installE}
\alias{installE}
\title{install package from my local drive}
\description{Just a shortcut to \code{\link{installB}} with the default \code{package="extremeStat"}}
\usage{installE(path="S:/Dropbox/Public")}
\arguments{
  \item{path}{path containing package folder. DEFAULT: "S:/Dropbox/Public")}
}
\author{Berry Boessenkool, \email{berry-b@gmx.de}, June 2015}
\seealso{\code{\link{installE}}, \code{\link{updatePackage}} }
