% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInstalled.R
\name{isInstalled}
\alias{isInstalled}
\title{Is a package installed and usable?}
\usage{
isInstalled(package)
}
\arguments{
\item{package}{Package character string passed to  \code{\link{requireNamespace}}}
}
\description{
Is a package installed and usable?
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June 2017
}
