% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndMessage.R
\name{loadAndMessage}
\alias{loadAndMessage}
\title{install package from local drive}
\usage{
loadAndMessage(package, quiet = TRUE)
}
\arguments{
\item{package}{Package name}

\item{quiet}{Logical: suppress messages like "package was built under R version xyz"}
}
\description{
load package and give verbose message
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2016
}
\seealso{
\code{\link{loadPackages}}
}
\keyword{package}

