% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{Load a number of packages I always like to have in the search path.}
\usage{
loadPackages()
}
\description{
Load a number of packages I always like to have in the search path.
}
