% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{newsdesc}
\alias{newsdesc}
\title{open NEWS and DESCRIPTION files at package root}
\usage{
newsdesc(path = ".")
}
\arguments{
\item{path}{Package dir. DEFAULT: "."}
}
\description{
open NEWS and DESCRIPTION files at package root in default viewer
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2019
}
