% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{newsfile}
\alias{newsfile}
\title{open NEWS file}
\usage{
newsfile(path = "NEWS")
}
\arguments{
\item{path}{Filename at pacakge root directory to be opened}
}
\description{
open NEWS file in default viewer
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2019
}
