% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{packagePath}
\alias{packagePath}
\title{Base path of package at current getwd}
\usage{
packagePath(path = ".")
}
\arguments{
\item{path}{Package path, Character string. DEFAULT: "." (current getwd)}
}
\description{
Base path of package at current getwd
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
