% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{packsNewR}
\alias{packsNewR}
\title{Packages in new R installation}
\usage{
packsNewR()
}
\description{
Packages in new R installation
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June 2017
}
