% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFinder.R
\name{pathFinder}
\alias{pathFinder}
\title{Find package path based on the computer used.}
\usage{
pathFinder(path = "C:/Dropbox/Rpack")
}
\arguments{
\item{path}{Path containing package folder. DEFAULT: "C:/Dropbox/Rpack"}
}
\description{
Find package path based on the computer used.
}
