% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackage.R
\name{updatePackage}
\alias{updatePackage}
\title{install package from local drive}
\usage{
updatePackage(package = "berryFunctions", path = "S:/Dropbox/Public")
}
\arguments{
\item{package}{Package name. DEFAULT: "berryFunctions"}

\item{path}{Path containing package folder. DEFAULT: "S:/Dropbox/Public"}
}
\value{
Nothing
}
\description{
install current development version of package from folder
}
\details{
Calls \code{\link{installB}} if the package date or version in the source code has been changed.
}
\examples{
# See ?loadPackages

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2014
}
\seealso{
\code{\link{installB}}, \code{\link{packageDescription}}, \code{\link{read.dcf}}
}
\keyword{package}

