% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{allFunctions}
\alias{allFunctions}
\title{create and open a file with the content of all the functions in all my packages}
\usage{
allFunctions(package = berryFunctions::packagePath(), ...)
}
\arguments{
\item{package}{Name(s) of packages. If NA, all in my package folder. 
DEFAULT: \code{berryFunctions::\link{packagePath}()}}

\item{\dots}{Arguments passed to \code{combineFiles} like quiet=TRUE}
}
\value{
file path
}
\description{
create and open a file with the content of all the functions in all my packages
}
\examples{
# allFunctions()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Okt 2016
}
