% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{course}
\alias{course}
\title{open R course PDF}
\usage{
course(pres = FALSE)
}
\arguments{
\item{pres}{Logical: Open presentation file with LaTeX onslide et al? DEFAULT: FALSE}
}
\description{
open R course PDF in default viewer
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2019
}
