% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{funInCode}
\alias{funInCode}
\title{Is a function used in some code?}
\usage{
funInCode(funname, code)
}
\arguments{
\item{funname}{Function name (character string)}

\item{code}{Code (character string)}
}
\description{
Report whether a function name is used in some code.
             Only returns TRUE if it is followed by "(" or included as pack::fun
}
\examples{
code <- "mean(pi); head(data.frame(A=1:5)); data_frame(7:9); do.call(base::print, list(pi))"
funs <- c("mean", "head", "head.data.frame",  "data.frame", "frame", "data_frame", "base::print")
data.frame(isfound=sapply(funs, funInCode, code=code), 
           shouldbefound=c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE, TRUE))

}
\seealso{
\url{https://stackoverflow.com/questions/45691254}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
