% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installB.R
\name{installB}
\alias{installB}
\title{install packages from local drive}
\usage{
installB(
  package = NA,
  path = "C:/Dropbox/Rpack",
  force = FALSE,
  load = TRUE,
  unloadrevdep = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{package}{Package name. DEFAULT: NA (interactive selection)}

\item{path}{Path containing package folder. DEFAULT: "C:/Dropbox/Rpack"}

\item{force}{Logical. Even install if the version is not outdated? DEFAULT: FALSE}

\item{load}{Logical. Also call loadAndMessage? DEFAULT: TRUE}

\item{unloadrevdep}{Try to unload some common reverse dependencies? DEFAULT: TRUE}

\item{quiet}{Suppress messages? DEFAULT: FALSE}

\item{\ldots}{passed to \code{devtools::\link[devtools]{install}}}
}
\description{
Remove function objects from workspace and try to unload reverse dependencies. 
Then call \code{devtools::\link[devtools]{install}}.\cr
}
\examples{
\dontrun{
# Here's what you could write into your Rprofile.site:

# Loading Packages
try(br::loadPackages())

options(help_type="html")

# Set CRAN repository
local({r <- getOption("repos")
r["CRAN"] <- "https://cran.rstudio.com" # 0-cloud
options(repos=r)})

# create character string  base::desktop
desktop <- "C:/Users/berry/Desktop" # linux "/home/berry/Desktop"

# Speed up package installation
# https://blog.jumpingrivers.com/posts/2017/speed_package_installation/
options(Ncpus = parallel::detectCores()-1 )

# if install.packages and download.files give HTTP status 403 forbidden:
options(url.method="libcurl")

} # end dontrun

}
\seealso{
\code{\link{packageDescription}}, \code{\link{read.dcf}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2014 - 2023
}
\keyword{package}
