% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOutdated.R, R/installA.R, R/installB.R,
%   R/loadAndMessage.R, R/loadPackages.R, R/pathFinder.R
\name{checkOutdated}
\alias{checkOutdated}
\alias{installA}
\alias{installB}
\alias{loadAndMessage}
\alias{loadPackages}
\alias{pathFinder}
\title{install packages from local drive}
\usage{
checkOutdated(package, path = "C:/Dropbox/Rpack", quiet = FALSE)

installA(path = "C:/Dropbox/Rpack", quiet = TRUE, ...)

installB(
  package = NA,
  path = "C:/Dropbox/Rpack",
  force = FALSE,
  load = TRUE,
  unloadrevdep = TRUE,
  quiet = FALSE,
  ...
)

loadAndMessage(package, quiet = TRUE, message = TRUE)

loadPackages()

pathFinder(path = "C:/Dropbox/Rpack")
}
\arguments{
\item{package}{Package name. DEFAULT: NA (interactive selection)}

\item{path}{Path containing package folder. DEFAULT: "C:/Dropbox/Rpack"}

\item{quiet}{Logical for loadAndMessage: suppress messages like "package was built under R version xyz" in loadAndMessage}

\item{force}{Logical. Even install if the version is not outdated? DEFAULT: FALSE}

\item{load}{Logical. Also call loadAndMessage? DEFAULT: TRUE}

\item{unloadrevdep}{Try to unload some common reverse dependencies? DEFAULT: TRUE}

\item{message}{Write custom message in loadAndMessage? DEFAULT: TRUE}

\item{\ldots}{Optional for installE and isntallO: path argument passed to installB}
}
\description{
\bold{installB} removes function objects from workspace and tries to unload
      reverse dependencies. It then calls \code{devtools::\link[devtools]{install}}.\cr
\bold{installA} runs this for all my packages.\cr
\bold{pathFinder} changes the path based on the computer used.\cr
\bold{loadAndMessage} calls \code{\link{require}} and gives verbose output.\cr
\bold{loadPackages} loads a number of packages I always like to have in the search path.\cr
}
\examples{
\dontrun{
# Here's what you could write into your Rprofile.site:

options(help_type="html")

# Set CRAN repository
local({r <- getOption("repos")
r["CRAN"] <- "https://cran.rstudio.com" # 0-cloud
options(repos=r)})

# create character string  base::desktop
desktop <- "C:/Users/berry/Desktop" # linux "/home/berry/Desktop"

# Speed up package installation
# https://blog.jumpingrivers.com/posts/2017/speed_package_installation/
options(Ncpus = parallel::detectCores()-1 )

# Loading Packages
try(br::loadPackages())

# if install.packages and download.files give HTTP status 403 forbidden:
options(url.method="libcurl")

} # end dontrun

}
\seealso{
\code{\link{packageDescription}}, \code{\link{read.dcf}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2014 - Nov 2016
}
\keyword{package}
