% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installE.R
\name{installE}
\alias{installE}
\title{install package from local drive}
\usage{
installE(path = "S:/Dropbox/Public")
}
\arguments{
\item{path}{Path containing package folder. DEFAULT: "S:/Dropbox/Public"}
}
\description{
Just a shortcut to \code{\link{installB}} with the default \code{package="extremeStat"}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June 2015
}
\seealso{
\code{\link{installE}}, \code{\link{updatePackage}}
}
\keyword{package}

