\name{installE}
\alias{installE}
\title{install package from my local drive}
\description{Just a shortcut to \code{\link{installB}} with different defaults}
\usage{installE(package="extremeStat", path="S:/Dropbox/Public")}
\arguments{
  \item{package}{package name. DEFAULT: "extremeStat"}
  \item{path}{path containing package folder. DEFAULT: "S:/Dropbox/Public")}
}
\author{Berry Boessenkool, \email{berry-b@gmx.de}, June 2015}
\seealso{\code{\link{installE}}, \code{\link{updatePackage}} }
