% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndMessage.R
\name{loadAndMessage}
\alias{loadAndMessage}
\title{Call \code{\link{require}} and give verbose output}
\usage{
loadAndMessage(package, quiet = TRUE, message = TRUE)
}
\arguments{
\item{package}{Package name}

\item{quiet}{suppress messages like "package was built under R version xyz". DEFAULT: TRUE}

\item{message}{Write custom message? DEFAULT: TRUE}
}
\description{
Call \code{\link{require}} and give verbose output
}
