% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{install package from local drive}
\usage{
loadPackages(ask = TRUE)
}
\arguments{
\item{ask}{Logical. Prompt for input? If FALSE, loadPackages acts as if input is 2. DEFAULT: TRUE}
}
\description{
load packages I always like to have in the search path, update my own packages
}
\examples{

\dontrun{
# Here's what you could write into your Rprofile.site:

options(help_type="html")

# Set CRAN repository
local({r <- getOption("repos")
r["CRAN"] <- "https://cran.rstudio.com" # 0-cloud
options(repos=r)})

desktop <- "C:/Users/berry/Desktop"

grDevices::windows.options(width=5.5, height=5)

.trPaths <- paste(Sys.getenv('LOCALAPPDATA'),
'\\\\Temp\\\\Tinn-R', c('', 'search.txt', 'objects.txt',
'file.r', 'selection.r', 'block.r','lines.r'),sep='\\\\')

# Loading Packages
if(interactive()) installB::loadPackages(ask=TRUE)


# Here's what I have in my Rconsole file:
# Style This can be `yes' (for MDI) or `no' (for SDI).
MDI = no
# Dimensions (in characters) of the console.
rows = 15
columns = 100
# Dimensions (in characters) of the internal pager.
pgrows = 15
pgcolumns = 100
# Initial position of the console (pixels, relative to the workspace for MDI)
xconsole = 1088
yconsole = 716    # or 840
# Initial position of the graphics window
xgraphics = -5
ygraphics = 0
} # end dontrun

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2016
}
\seealso{
\code{\link{updatePackage}}, \code{\link{installB}}
}
\keyword{package}

