% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromDevtools.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{document}
\alias{check}
\alias{install}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{devtools}{\code{\link[devtools]{check}}, \code{\link[devtools]{document}}, \code{\link[devtools]{install}}}
}}

